third_party/blink/web_tests/wpt_internal/prerender/session-history-activation.https.html

<!DOCTYPE html>
<title>Test history.length</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/session-history-test-util.js"></script>
<body>
  <script>
    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testHistoryLengthInPrerender"),
        "Passed",
        "test in prerender"
      );
      assert_equals(
        await runTestInActivatedPage("testHistoryLengthInPrerender"),
        "Passed",
        "test in activated page"
      );
    }, "history.length should be updated after activation");
  </script>
</body>
